\name{mvblasso-package}
\alias{mvblasso-package}
\alias{mvblasso}
\docType{package}
\title{
\packageTitle{mvblasso}
}
\description{
\packageDescription{mvblasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mvblasso}
\packageIndices{mvblasso}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mvblasso}

Maintainer: \packageMaintainer{mvblasso}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
